add quickview tests; refactoring
